+++ BattleNetwork/bnMiscBackground.cpp
@@ -25,8 +25,8 @@ void MiscBackground::Update(double _elapsed) {
 
   animation.Update(_elapsed, dummy);
 
-  y -= 0.5f * _elapsed;
-  x -= 0.5f *_elapsed;
+  y -= static_cast<float>(0.5f * _elapsed);
+  x -= static_cast<float>(0.5f *_elapsed);
 
   if (x < 0) x = 1;
   if (y < 0) y = 1;
