+++ BattleNetwork/bnSharedHitbox.h
@@ -52,7 +52,7 @@ public:
   const float GetHeight() const override;
   
 private:
-  float cooldown; /*< When cooldown reaches zero, this hitbox removes */
+  double cooldown; /*< When cooldown reaches zero, this hitbox removes */
   bool keepAlive; /*!< If duration is not set, the hitbox stays alive for long as the owner stays alive*/
   Spell* owner; /*!< When this hitbox is hit, the owner is hit */
 };
