+++ BattleNetwork/bnUndernetBackground.h
@@ -21,9 +21,9 @@ public:
   void Update(double _elapsed) override;
 
 private:
-  float progress; /**< Animation progress */
+  double progress; /**< Animation progress */
   int colorIndex; /**< The current color to flash */
   std::vector<sf::Color> colors; /**< In the game, the undernet flashes colors */
-  float colorProgress; /**< Counter until next flash */
+  double colorProgress; /**< Counter until next flash */
   sf::Time colorDuration; /**< Duration inbetween color clashes */
 };
